pyproject.toml
[project]
name = "ethics-council"
version = "0.1.0"
description = "Ethics Council — Multi-agent ethics review roundtable forum (configurable framework)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "jsonschema>=4.0",
]

[project.scripts]
ethics-council = "engine.cli:main"

[tool.setuptools.packages.find]
include = ["engine*", "config*", "backend*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
